+++ lib/wechat/core.rb
@@ -1,5 +1,11 @@
 require 'wechat/core/version'
 
+require 'wechat/core/common'
+require 'wechat/core/access_token'
+
+require 'wechat/core/server_address'
+require 'wechat/core/tiny_link'
+
 module Wechat
   module Core
     # Your code goes here...